Allow Demonstration.delete() to call community.delete()

+++ demo/models.py
@@ -36,7 +36,7 @@ class Demonstration(models.Model):
         if self.log is not None:
             self.log.deleted_at = datetime.datetime.utcnow()
             self.log.save()
-        if self.community is not None:
+        if self.community is not None and self.community.id is not None:
             self.community.delete()
         return super(Demonstration, self).delete(*args, **kwargs)
